// src/common/enums/item.enum.ts
// src/common/enums/item.enum.ts
export enum ItemType {
  ITEM = 'ITEM',
  SERVICE = 'Service',
  RAW_MATERIAL = 'Raw Material',
  CONSUMER_PRODUCT = 'Consumer Product',
  FRAGRANCES_EO = 'Fragrances  EO',
  HOUSE_KEEPING = 'House Keeping',
  PACKING_MATERIAL = 'Packing Material',
  CUSTOMIZED_RANGE = 'Customized Range',
  STOCKS_WIP = 'Stocks of Work In Progress',
  FURNITURE_EQUIPMENT = 'Furniture and Equipment',
  MACHINERY_EQUIPMENT = 'Machinaery and Equipment',
  LAB_MACHINERY = 'Lab Machinery and Equpment',
  COMPUTER_EQUIPMENT = 'Computer and Equipments',
  FACTORY_BUILDING = 'Factory Bulding',
}

export enum MBFlag {
  B = 'B', // Buy
  M = 'M', // Manufacture
  D = 'D', // Service/Direct
  F = 'F', // Fixed Asset
}

export enum UnitType {
  PCS = 'pcs',
  KG = 'Kg',
  LTR = 'ltr',
  BOXES = 'boxes',
  NOS = 'Nos',
}

export enum ItemCategory {
  HOREKA_RANGE = 'Horeka Range',
  RAW_MATERIAL = 'Raw Material',
  SERVICE = 'Service',
  CONSUMER_PRODUCT = 'Consumer Product',
  FRAGRANCES_EO = 'Fragrances  EO',
  HOUSE_KEEPING = 'House Keeping',
  PACKING_MATERIAL = 'Packing Material',
  CUSTOMIZED_RANGE = 'Customized Range',
  STOCKS_WIP = 'Stocks of Work In Progress',
}